+++ packages/tools/src/drawingSvg/drawEllipse.ts
@@ -12,20 +12,20 @@ function drawEllipse(
   corner2: Types.Point2,
   options = {},
   dataId = ''
-){
-  const top: Types.Point2 = [ (corner1[0] + corner2[0]) / 2, corner1[1] ];
-  const bottom: Types.Point2 = [ (corner1[0] + corner2[0]) / 2, corner2[1] ];
-  const left: Types.Point2 = [ corner1[0], (corner1[1] + corner2[1]) / 2];
-  const right: Types.Point2 = [ corner2[0], (corner1[1] + corner2[1]) / 2];
+) {
+  const top: Types.Point2 = [(corner1[0] + corner2[0]) / 2, corner1[1]];
+  const bottom: Types.Point2 = [(corner1[0] + corner2[0]) / 2, corner2[1]];
+  const left: Types.Point2 = [corner1[0], (corner1[1] + corner2[1]) / 2];
+  const right: Types.Point2 = [corner2[0], (corner1[1] + corner2[1]) / 2];
 
   drawEllipseByCoordinates(
     svgDrawingHelper,
     annotationUID,
     ellipseUID,
     [bottom, top, left, right],
-    options = {},
-    dataId = ''
-  )
+    (options = {}),
+    (dataId = '')
+  );
 }
 
 export default drawEllipse;