+++ packages/tools/src/enums/index.js
@@ -4,6 +4,7 @@ import AnnotationStyleStates from './AnnotationStyleStates';
 import Events from './Events';
 import SegmentationRepresentations from './SegmentationRepresentations';
 import { Swipe } from './Touch';
+import StrategyCallbacks from './StrategyCallbacks';
 
 export {
   MouseBindings,
@@ -13,4 +14,5 @@ export {
   Events,
   SegmentationRepresentations,
   Swipe,
+  StrategyCallbacks,
 };