+++ packages/tools/src/types/FloodFillTypes.ts
@@ -5,13 +5,13 @@ type FloodFillResult = {
   boundaries: Types.Point2[] | Types.Point3[];
 };
 
-type FloodFillGetter3D = (x: number, y: number, z: number) => number;
-type FloodFillGetter2D = (x: number, y: number) => number;
+type FloodFillGetter3D = (x: number, y: number, z: number) => unknown;
+type FloodFillGetter2D = (x: number, y: number) => unknown;
 type FloodFillGetter = FloodFillGetter2D | FloodFillGetter3D;
 
 type FloodFillOptions = {
-  onFlood?: (x, y) => void;
-  onBoundary?: (x, y) => void;
+  onFlood?: (x: number, y: number, z?: number) => void;
+  onBoundary?: (x: number, y: number, z?: number) => void;
   equals?: (a, b) => boolean; // Equality operation for your datastructure. Defaults to a === b.
   diagonals?: boolean; // Whether to flood fill across diagonals. Default false.
 };